Demonstration of scaling parameters
For the request of reviewers we added a small demonstration of how environmental and phenotypic parameters can be scaled. In the original model the environmental value was between 0.9 and 1.1 in the metapopulation's two extremes. We can scale it up to 9 and 11 by simultaneously transforming the phenotype values and the width of the selection function. We can also re-center it to, for example 24-26 (which is a reasonable mean temperature range for the Great Barrier Reef for example).