Test of heterozygosity and genetic variance
This folder has been added in response to Reviewer 1. Here we provide modified SLiM models, shell scripts and R scripts to demonstrate the influence of mutation effect and the number of genes on heterozygosity and additive genetic variance. If you chose not to re-run the models, and would like to have the data files to run the R scripts, please drop us a line.